pyproject.toml
[project]
name = "python-advanced-gui-programming"
version = "0.1.0"
authors = [
    {name = "Piotr Opała", email = "opalapj@gmail.com"},
]
license = {text = "MIT"}
dependencies = ["Pillow>=12.0.0"]
requires-python = "==3.13.*"

[tool.pdm]
distribution = false
